[REFACTOR] Split InvoiceMapperService tests into separate files
- Moved all InvoiceMapperService tests into a new dedicated folder `tests/services/InvoiceMapperService/`.
- Separated tests into individual files for better organization and maintainability.
