+++ backend/tests/mocks/models.mock.js
@@ -1,0 +1,6 @@
+jest.mock('../../../src/models', () => ({
+  Invoice: {
+    build: jest.fn().mockImplementation(data => data)
+  },
+  sequelize: {}
+}));
